pyproject.toml
[project]
name = "superforecaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.7.0",
    "httpx>=0.24.0",
    "logfire>=4.25.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.1.0",
    "python-dotenv>=1.0.0",
]
